Bug fix.  We were not making a proper unique_index for the umts layer 3 info.  All cells were treated the same which would disrupt the segmented sib portion of the decoding.

+++ src/rf_phreaker/layer_3_common/bcch_bch_message_aggregate.h
@@ -15,7 +15,7 @@ typedef uint16_string mcc_type;
 typedef uint16_string mnc_type;
 typedef uint16_t lac_type;
 typedef uint32_t cid_type;
-typedef uint32_t unique_sector_key_type;
+typedef int64_t unique_sector_key_type;
 
 class bcch_bch_message_aggregate
 {
